Fix rendering of embedded objects in CompositionScanner
In the past, doing certain™ things in the babylonian browser frequently crashed it to crash for me with awkward rendering errors. Skimming through the existing stop conditions on the composition scanner revealed that all of them set lastIndex in the TextLine iff they return true (and thus break the loop in #composeFrom:inRectangle:firstLine:leftSide:rightSide:). With this change, my special example does not break the rendering engine any longer:

	foo: aNumber with: anotherNumber
	<exampleNamed: 'some example 5' self: 'TestObject new' with: '2' with: '8'>
	| tmp |
	tmp := aNumber * anotherNumber.
	^ "<bpProbe id: 4248111>""<bpReplace for: 'all examples' with: []>"tmp"</bpReplace>""</bpProbe>" sqrt